Icamera リファクタリング

+++ ICamera.cpp
@@ -1,6 +1,6 @@
 #include "ICamera.h"
 
-void ICamera::SetProjection(float fovY, float nearZ, float farZ, float width_, float height_)
+void ICamera::SetProjection(float fovY, float nearZ, float farZ, int width_, int height_)
 {
 	matProjection = MyMath::PerspectiveFovLH(width_, height_, fovY, nearZ, farZ);
 }
